pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "mistralai>=2.4.0",
    "psycopg2-binary>=2.9.12",
]
